+++ packages/identity/CHANGELOG.md
@@ -3,9 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# [4.0.0](https://github.com/ZbayApp/identity/compare/@zbayapp/identity@4.0.0-alpha.0...@zbayapp/identity@4.0.0) (2021-12-29)
+# [0.1.0-alpha.1](https://github.com/ZbayApp/identity/compare/@quiet/identity@0.1.0-alpha.0...@quiet/identity@0.1.0-alpha.1) (2022-01-27)
 
-**Note:** Version bump only for package @zbayapp/identity
+**Note:** Version bump only for package @quiet/identity
+
+
+
+
+
+# [0.1.0-alpha.0](https://github.com/ZbayApp/identity/compare/@quiet/identity@5.1.0-alpha.0...@quiet/identity@0.1.0-alpha.0) (2022-01-25)
+
+**Note:** Version bump only for package @quiet/identity
+
+
+
+
+
+# 5.1.0-alpha.0 (2022-01-25)
+
+**Note:** Version bump only for package @quiet/identity
+
+
+
+
+
+# [5.0.0-alpha.0](https://github.com/quiet/identity/compare/@quiet/identity@4.0.0...@quiet/identity@5.0.0-alpha.0) (2022-01-24)
+
+**Note:** Version bump only for package @quiet/identity
+
+
+
+
+
+# [4.0.0](https://github.com/quiet/identity/compare/@quiet/identity@4.0.0-alpha.0...@quiet/identity@4.0.0) (2021-12-29)
+
+**Note:** Version bump only for package @quiet/identity
 
 
 
@@ -13,4 +45,4 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # 4.0.0-alpha.0 (2021-12-28)
 
-**Note:** Version bump only for package @zbayapp/identity
+**Note:** Version bump only for package @quiet/identity